tests/wpt/tests/css/css-color-hdr/parsing.html

<!DOCTYPE html>
<title>Parsing dynamic-range-limit with valid and invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-color-hdr/#the-dynamic-range-limit-property">
<meta name="assert" content="Parsing of dynamic-range-limit">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
  test_valid_value("dynamic-range-limit", "standard");
  test_valid_value("dynamic-range-limit", "high");
  test_valid_value("dynamic-range-limit", "constrained-high");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(high 80%, standard 20%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(high 8%, standard 2%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(high 99%, standard 99%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(high 99%, standard 99%, constrained-high 10%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(high 1%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(constrained-high 20%, high 80%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(dynamic-range-limit-mix(constrained-high 90%, high 10%) 1%, high 80%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(high 1%, dynamic-range-limit-mix(constrained-high 2%, high 10%) 80%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(dynamic-range-limit-mix(high 1%, standard 2%) 3%, dynamic-range-limit-mix(constrained-high 5%, high 5%) 6%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(high 80%, standard 20%, )");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(high, standard 20%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(constrained-high, high, 80%)");
  test_invalid_value("dynamic-range-limit", "");
  test_invalid_value("dynamic-range-limit", "none");
  test_invalid_value("dynamic-range-limit", "default");
  test_invalid_value("dynamic-range-limit", "hdr");
  test_invalid_value("dynamic-range-limit", "sdr");
  test_invalid_value("dynamic-range-limit", "low");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(high 80% standard 20%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(low, high, 10%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(high 101%, standard 1%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(high -1%, standard 1%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(standard, high, 0.1)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(high 0%, standard 0%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(dynamic-range-limit-mix(high 1%, standard 2%) 3%, dynamic-range-limit-mix(constrained-high 0%, high 0%) 6%)");
</script>
